+++ test/dummy/app/models/message_root.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class MessageRoot < ApplicationRecord
+  recording_studio_recordable label: "Messages", root: true, shared: true
+
+  validates :name, presence: true
+end